settings/ui: fix AI settings persistence and macOS conversion warning
- AiProviderSettings::save/restore were prefixing 'aiProviders/' again while
  the caller in Configuration already scoped into that group, leading to
  /aiProviders/aiProviders/<provider> keys. Drop the redundant prefix.
- Cast the result of QStringList::indexOf to int to silence -Wshorten-64-to-32
  on 64-bit macOS builds.

+++ src/ui/src/aiclibackend.cpp
@@ -344,9 +344,8 @@ bool AiCliBackend::start( Provider provider, const QString& prompt, const QByteA
         sandboxArgs << providerArgs;
 
         if ( provider == Provider::CodeBuddy && QFileInfo( userState ).isFile() ) {
-            const int insertAt
-                = sandboxArgs.indexOf( QStringLiteral( "--dir" ),
-                                       sandboxArgs.indexOf( authDir ) );
+            const int insertAt = static_cast<int>( sandboxArgs.indexOf(
+                QStringLiteral( "--dir" ), sandboxArgs.indexOf( authDir ) ) );
             if ( insertAt >= 0 ) {
                 sandboxArgs.insert( insertAt, QStringLiteral( "--ro-bind" ) );
                 sandboxArgs.insert( insertAt + 1, userState );
